BUG: Camera In Floor
So I've opted to try having an object follow the player from a distance (like the pikmin manager), and the Camera will move towards that position. That way I can control the camera acceleataion and stuff more easily with the camera detached from the player. Need to figure out though why the CameraStateMachine isnt attracting the camera on the Y-Axis